+++ lib/extra/style.dart
@@ -5,20 +5,17 @@ import 'package:date_picker_timeline/extra/dimen.dart';
 const TextStyle defaultMonthTextStyle = TextStyle(
   color: AppColors.defaultMonthColor,
   fontSize: Dimen.monthTextSize,
-  fontFamily: 'Roboto',
   fontWeight: FontWeight.w500,
 );
 
 const TextStyle defaultDateTextStyle = TextStyle(
   color: AppColors.defaultDateColor,
   fontSize: Dimen.dateTextSize,
-  fontFamily: 'Roboto',
   fontWeight: FontWeight.w500,
 );
 
 const TextStyle defaultDayTextStyle = TextStyle(
   color: AppColors.defaultDayColor,
   fontSize: Dimen.dayTextSize,
-  fontFamily: 'Roboto',
   fontWeight: FontWeight.w500,
 );